// fe/src/lib/api/generated/model/graduateTracerDistributionSection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { GraduateTracerDistributionRow } from './graduateTracerDistributionRow';

export interface GraduateTracerDistributionSection {
  denominator: number;
  denominator_label: string;
  key: string;
  label: string;
  multiple_selection: boolean;
  rows: GraduateTracerDistributionRow[];
}